\name{multiclassLR-package}
\alias{multiclassLR-package}
\alias{multiclassLR}
\docType{package}
\title{
\packageTitle{multiclassLR}
}
\description{
\packageDescription{multiclassLR}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{multiclassLR}
\packageIndices{multiclassLR}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{multiclassLR}

Maintainer: \packageMaintainer{multiclassLR}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
